requirements.txt
requests~=2.31.0
selenium~=4.20.0
pillow~=10.3.0